test(subscription_service): fix BDD steps after interaction-event refactor

+++ feed_service/src/main/java/com/kolmir/feed_service/repository/CommentRepository.java
@@ -12,7 +12,7 @@ import com.kolmir.feed_service.model.Comment;
 @Repository
 public interface CommentRepository extends JpaRepository<Comment, Long> {
     public Page<Comment> findAllByPostId(Long postId, Pageable pageable);
-    public int findCountByPostId(Long postId);
+    public long findCountByPostId(Long postId);
     public Page<Comment> findAllByAuthorIdAndPostId(Long authorId, Long postId, Pageable pageable);
     public boolean existsByIdAndAuthorId(Long commentId, Long postId);
 }
